ENH: add tab with the beamclass table
